Made unit testing work
