+++ resumator-ui/package.json
@@ -1,6 +1,10 @@
 {
   "name": "resumator",
   "version": "1.0.0",
+  "scripts": {
+    "start": "gulp develop",
+    "test": "gulp runTests"
+  },
   "dependencies": {
     "babel-polyfill": "^6.3.14",
     "body-parser": "^1.14.2",
@@ -74,6 +78,7 @@
     "gulp-util": "^3.0.7",
     "imagemin-pngquant": "^4.2.0",
     "less-plugin-autoprefix": "^1.5.1",
+    "plop": "^1.5.0",
     "react-addons-test-utils": "^0.14.7",
     "rimraf": "^2.5.0",
     "run-sequence": "^1.1.5",
