--- resumator-ui/test/TestTest.js
@@ -1,12 +1,0 @@
-//
-// import React from 'react';
-// import { mount, shallow } from 'enzyme';
-//
-// describe('<Foo />', () => {
-//
-//   it('calls componentDidMount', () => {
-//     const wrapper = mount(<Foo />);
-//     expect(Foo.prototype.componentDidMount.calledOnce).to.equal(true);
-//   });
-//
-// });
