--- resumator-ui/withDom.js
@@ -1,22 +1,0 @@
-if (!global.document) {
-  try {
-    const jsdom = require('jsdom').jsdom; // could throw
-
-    const exposedProperties = ['window', 'navigator', 'document'];
-
-    global.document = jsdom('');
-    global.window = document.defaultView;
-    Object.keys(document.defaultView).forEach((property) => {
-      if (typeof global[property] === 'undefined') {
-        exposedProperties.push(property);
-        global[property] = document.defaultView[property];
-      }
-    });
-
-    global.navigator = {
-      userAgent: 'node.js',
-    };
-  } catch (e) {
-    // jsdom is not supported...
-  }
-}
